A7_code/requirements.txt
fastapi
uvicorn
numpy
pillow
prometheus-fastapi-instrumentator
argparse
scipy
tensorflow